feat: backup provider (#33)
This PR adds a backup provider, a job manager, runner and job/backup storages.

The `BackupProvider` provides access to :

* Backup `jobs` (a reactive list of backups that are currently in progress)
* Completed `backups`(a reactive list of completed jobs)
* A function that allows a new backup to be scheduled `addBackupJob`

There's a `JobStorage` which stores details of _pending_ backup jobs (i.e. they are queued, running or failed). You can add, update, list and remove from this store. The implementation in the PR is backed by `localStorage` (as the jobs are currently run locally it makes sense for this to be local right now).

There's a `BackupStorage` which stores details of successful backups. This is designed to be persisted across devices (so you can see your backups on your phone/desktop). It is backed by `cloudStorage` currently but ideally we'd swap this out for UCN and Storacha.

There's a `JobManager` which is in charge of scheduling and running jobs, and updating the 2 stores when things change (you submit your jobs to the manager).