fix: fetch high quality images for profile pictures in the list of chats
It turns out the only way to get high quality images is
to request them from the API. Furthermore, not all images are loadable using `client.getEntity` but it's difficult to determine which will work and which won't.

This PR introduces two things:

1) some client-side code to attempt to load the high quality version of an image using a new API endpoint in this app
2) the new API endpoint itself, which tries to fetch the image a couple different ways and then sets cache headers that tell the browser to consider the image immutable

The net result here is that we'll (a) try pretty hard to load images from Telegram and then (b) aggressively cache successful results - taken alone (a) might be a cause for performance concern, but (b) should compensate pretty well for that.