fix: get build working
Something about importing the server components from the route caused very weird errors - TypeScript was showing types different from the types that were imported in a file we haven't touched. I'm pretty convinced this is a deeper bug related to importing the vendor'd gram.js and the npm gram.js into the same environment, but that's pretty weird.

I'd like to track down the core issue eventually now but this fixes the build!