feat: after one failed login attempt, give the user a way to reset login state
We're seeing an error in production (https://github.com/storacha/project-tracking/issues/539) that has proven difficult to reproduce on a development machine, so this PR introduces an attempt to allow users to reset their auth state and log back in.

I'm not 100% sure this will actually solve the problem for the user - I know clearing local and session state entirely will work, and I know that isn't done by the telegram logout function, but I want to see if this fixes the issue before expanding the scope of what's cleared - in particular I'd like to avoid clearing local storage if possible, since you might have more than one user's information in it. if simply logging out of telegram allows the user to log in successfully, I think this is enough of a fix for now.